refactor(kinematics): update_odom 时间参数对齐 uint64_t, 消除窄化
- 原 uint16_t 上限 65535ms, 超 65 秒即截断
- 与 update_motor_speed 的 dt (uint64_t) 类型统一

+++ lib/Kinematics/Kinematics.cpp
@@ -82,7 +82,7 @@ void Kinematics::update_motor_speed(uint64_t now, const int32_t ticks[2]) {
 
 float Kinematics::get_motor_speed(uint8_t motor_id) { return current_motor_speeds_[motor_id]; }
 
-void Kinematics::update_odom(uint16_t dt) {
+void Kinematics::update_odom(uint64_t dt) {
     // 单位换算, ms -> s
     float dt_s = static_cast<float>(dt) / MS_TO_S;
